framework-comparisons/03-rag-implementation/crewai/requirements.txt
crewai==1.9.1
python-dotenv==1.1.1
langchain-openai==0.2.14
langchain-community==0.3.13
faiss-cpu==1.9.0.post1
numpy==1.26.4
